Break out CampignFilter timeframe select to separate component
